# docs/swagger.yaml
basePath: /
definitions:
...
  industry_news.CreateIndustryNewsRequest:
    properties:
      authorId:
        type: integer
      categoryId:
        type: integer
      content:
        minLength: 100
        type: string
      excerpt:
        maxLength: 500
        minLength: 50
        type: string
      location:
        type: string
      metadata:
        $ref: '#/definitions/industry_news.IndustryNewsMetadata'
      publishDate:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/industry_news.IndustryNewsStatus'
        enum:
        - draft
        - review
        - published
      tags:
        type: string
      title:
        maxLength: 200
        minLength: 10
        type: string
    required:
    - authorId
    - categoryId
    - content
    - excerpt
    - publishDate
    - status
    - title
    type: object
...
  industry_news.IndustryNewsListResponse:
    properties:
      industryNews:
        items:
          $ref: '#/definitions/industry_news.IndustryNews'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  industry_news.IndustryNewsResponse:
    properties:
      industryNews:
        $ref: '#/definitions/industry_news.IndustryNews'
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: 192.168.1.2:8081
info:
  contact:
    email: support@example.com
    name: API Support
  description: RESTful API for healthcare market research reports and categories with
    caching and pagination support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Healthcare Market Research API
  version: "1.0"
paths:
...
  /api/v1/categories/{slug}/industry-news:
    get:
      consumes:
      - application/json
      description: Get a paginated list of published industry news articles for a
        specific category
      parameters:
      - description: Category slug
        in: path
        name: slug
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of industry news articles with pagination
          schema:
            $ref: '#/definitions/industry_news.IndustryNewsListResponse'
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Get industry news by category slug
      tags:
      - IndustryNews
...
  /api/v1/industry-news:
    get:
      consumes:
      - application/json
      description: Get a paginated list of industry news articles with optional filtering
      parameters:
      - description: 'Filter by status: draft, review, published'
        in: query
        name: status
        type: string
      - description: Filter by category ID
        in: query
        name: categoryId
        type: integer
      - description: Filter by category slug
        in: query
        name: category
        type: string
      - description: Filter by tags (comma-separated)
        in: query
        name: tags
        type: string
      - description: Filter by author ID
        in: query
        name: authorId
        type: integer
      - description: Filter by location
        in: query
        name: location
        type: string
      - description: Search in title, excerpt, content
        in: query
        name: search
        type: string
      - description: Set to 'true' to list soft-deleted (trashed) articles
        in: query
        name: deleted
        type: string
      - description: 'Sort order: publish_date_desc, created_at_desc, updated_at_desc'
        in: query
        name: sort_by
        type: string
      - description: 'Page number (default: 1, min: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of industry news articles with pagination
          schema:
            $ref: '#/definitions/industry_news.IndustryNewsListResponse'
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Get all industry news articles
      tags:
      - IndustryNews
    post:
      consumes:
      - application/json
      description: Create a new industry news article. Admin/editor only.
      parameters:
      - description: Industry news data
        in: body
        name: industryNews
        required: true
        schema:
          $ref: '#/definitions/industry_news.CreateIndustryNewsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Industry news article created successfully
          schema:
            $ref: '#/definitions/industry_news.IndustryNewsResponse'
        "400":
          description: Bad request - invalid input or validation error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create industry news article
      tags:
      - IndustryNews
...
  /api/v1/industry-news/{id}/cancel-schedule:
    patch:
      consumes:
      - application/json
      description: Cancel scheduled publishing for an industry news article. Admin/editor
        only.
      parameters:
      - description: Industry news article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Schedule cancelled
          schema:
            $ref: '#/definitions/industry_news.IndustryNewsResponse'
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Cancel scheduled publish
      tags:
      - IndustryNews
  /api/v1/industry-news/{id}/publish:
    patch:
      consumes:
      - application/json
      description: Change industry news article status to published. Admin/editor
        only.
      parameters:
      - description: Industry news article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Industry news article published successfully
          schema:
            $ref: '#/definitions/industry_news.IndustryNewsResponse'
        "400":
          description: Bad request - invalid ID or article cannot be published
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin or editor role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Industry news article not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Publish industry news article
      tags:
      - IndustryNews
  /api/v1/industry-news/{id}/restore:
    patch:
      consumes:
      - application/json
      description: Restore a soft deleted industry news article by ID. Admin only.
      parameters:
      - description: Industry news article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Industry news article restored successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request - invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden - requires admin role
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Restore industry news article
      tags:
      - IndustryNews
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...